some optimisations

+++ src/Dimmer32u4.h
@@ -10,21 +10,20 @@
 
 class Dimmer32u4
 {
-  public:
-    Dimmer32u4(int pin);
-    Dimmer32u4(int pin, bool invert);
-    void setValue(int value);
-	int getMaxValue();
-	int getResolution();
-	void init();
-  private:
-	int _maxValue;
-    int _pin;
-	int _frequency;
-	int _value;
-	int _resolution;
-	bool _invert;
-	
+    public:
+        Dimmer32u4(uint8_t pin);
+        Dimmer32u4(uint8_t pin, bool invert);
+        void setValue(uint16_t value);
+        uint16_t getMaxValue();
+        uint8_t  getResolution();
+        void init();
+    private:
+        uint16_t _maxValue;
+        uint8_t  _pin;
+        uint16_t _frequency;
+        uint16_t _value;
+        uint8_t  _resolution;
+        bool     _invert;
 
 };
